CQ-4301935 Rename searchnlister to searchlister
@review @rismehta @sakshi-arora-1
