+++ bundles/core/src/test/resources/searchlister/exporter-searchlister-v1.json
@@ -1,11 +1,11 @@
 
 {
-    "id"                    : "searchnlister-ff2359bd7e",
+    "id"                    : "searchlister-fb0b07e803",
     "title"                 : "Sample Title",
     "layout"                : "Custom",
     "advancedSearchDisabled": true,
     "textSearchDisabled"    : true,
     "resultLimit"           : 4,
     "searchResults"         : {"data":[],"success":"true","nextOffset":0},
-    ":type"                 : "core/fd/components/formsportal/searchnlister/v1/searchnlister"
+    ":type"                 : "core/fd/components/formsportal/searchlister/v1/searchlister"
 }
