+++ examples/ui.apps/src/main/content/jcr_root/apps/forms-components-examples/components/searchlister/.content.xml
@@ -3,5 +3,5 @@
     jcr:description="Search and Lister Component for the Core Components Library"
     jcr:primaryType="cq:Component"
     jcr:title="Search and Lister"
-    sling:resourceSuperType="core/fd/components/formsportal/searchnlister/v1/searchnlister"
+    sling:resourceSuperType="core/fd/components/formsportal/searchlister/v1/searchlister"
     componentGroup="Core Components Examples"/>
