+++ examples/ui.content/src/main/content/jcr_root/content/core-components-examples/library/forms-and-communications-portal/.content.xml
@@ -15,5 +15,5 @@
         hideInNav="true"
         redirectTarget="/content/core-components-examples/library"/>
     <link/>
-    <searchnlister/>
+    <searchlister/>
 </jcr:root>
