+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/formsportal/portallister/v1/portallister/_cq_dialog/.content.xml
@@ -17,7 +17,7 @@
 <jcr:root xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:granite="http://www.adobe.com/jcr/granite/1.0" xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0" xmlns:nt="http://www.jcp.org/jcr/nt/1.0"
           jcr:primaryType="nt:unstructured"
           jcr:title="Edit Portal Lister Component"
-          extraClientlibs="core.forms.components.formsportal.searchnlister.v1.editor"
+          extraClientlibs="core.forms.components.formsportal.searchlister.v1.editor"
           sling:resourceType="cq/gui/components/authoring/dialog">
     <content
         jcr:primaryType="nt:unstructured"
