+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/formsportal/searchlister/v1/.content.xml
@@ -14,9 +14,7 @@
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~-->
-<jcr:root xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:jcr="http://www.jcp.org/jcr/1.0"
-          jcr:primaryType="cq:ClientLibraryFolder"
-          allowProxy="{Boolean}true"
-          categories="[core.forms.components.formsportal.searchnlister.v1]"
-          embed="[core.forms.components.commons.v1.clientlibs.layouts.iteminjector]"
-          dependencies="[jquery]"/>
+<jcr:root xmlns:jcr="http://www.jcp.org/jcr/1.0" xmlns:nt="http://www.jcp.org/jcr/nt/1.0"
+          jcr:primaryType="nt:unstructured">
+    <searchlister/>
+</jcr:root>
