+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/formsportal/searchlister/v1/searchlister/clientlibs/site/css/search.less
@@ -14,10 +14,10 @@
  * limitations under the License.
  ******************************************************************************/
 
-.cmp-searchnlister-search {
+.cmp-searchlister-search {
     height: 40px;
 
-    .cmp-searchnlister-search__box {
+    .cmp-searchlister-search__box {
         position: relative;
         width: 35%;
         display: inline-block;
@@ -48,7 +48,7 @@
         }
     }
 
-    .cmp-searchnlister-search__sortbutton {
+    .cmp-searchlister-search__sortbutton {
         height: 40px;
         background-color: #FFFFFF;
         border: 1px solid #CFD8DC;
@@ -66,7 +66,7 @@
     }
 }
 
-.cmp-searchnlister__more {
+.cmp-searchlister__more {
     color: #455A64;
     font-size: 16px;
     display: inline-block;
@@ -85,7 +85,7 @@
     }
 }
 
-.cmp-searchnlister__more--wrapper {
+.cmp-searchlister__more--wrapper {
     text-align: center;
     display: block;
 }
