+++ ui.apps/src/main/content/jcr_root/apps/core/fd/components/formsportal/searchlister/v1/searchlister/searchlister.html
@@ -15,35 +15,35 @@
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/-->
 
 <sly data-sly-use.model="com.adobe.cq.forms.core.components.models.formsportal.SearchAndLister"
-     data-sly-use.logic="${'searchnlister.js' @ layout=model.layout}"
+     data-sly-use.logic="${'searchlister.js' @ layout=model.layout}"
      data-sly-use.clientlib="${'/libs/granite/sightly/templates/clientlib.html'}"
      data-sly-use.items="${logic.layoutPath}"></sly>
-<div class="cmp-searchnlister" data-cmp-is="formssearch" id="${model.id}" data-queryPath="${resource.path}.model.json" data-limit="${model.resultLimit}">
-    <h1 class="cmp-searchnlister__heading" data-sly-text="${model.title}">Forms Portal</h1>
-    <div class="cmp-searchnlister__seperator" data-sly-resource.seperator="${logic.seperator}"></div>
-    <div class="cmp-searchnlister-search" data-sly-test="${!model.textSearchDisabled || !model.advancedSearchDisabled}">
-        <div class="cmp-searchnlister-search__box" data-sly-test.textSearch="${!model.textSearchDisabled}">
-            <input class="cmp-searchnlister-search__box--text" data-cmp-hook-formssearch="input" type="text" placeholder="Search">
+<div class="cmp-searchlister" data-cmp-is="formssearch" id="${model.id}" data-queryPath="${resource.path}.model.json" data-limit="${model.resultLimit}">
+    <h1 class="cmp-searchlister__heading" data-sly-text="${model.title}">Forms Portal</h1>
+    <div class="cmp-searchlister__seperator" data-sly-resource.seperator="${logic.seperator}"></div>
+    <div class="cmp-searchlister-search" data-sly-test="${!model.textSearchDisabled || !model.advancedSearchDisabled}">
+        <div class="cmp-searchlister-search__box" data-sly-test.textSearch="${!model.textSearchDisabled}">
+            <input class="cmp-searchlister-search__box--text" data-cmp-hook-formssearch="input" type="text" placeholder="Search">
         </div>
-        <div class="cmp-searchnlister-search__sortbutton--wrapper" data-sly-test.advancedSearch="${!model.advancedSearchDisabled}">
-            <select class="cmp-searchnlister-search__sortbutton" data-cmp-hook-formssearch="filter">
+        <div class="cmp-searchlister-search__sortbutton--wrapper" data-sly-test.advancedSearch="${!model.advancedSearchDisabled}">
+            <select class="cmp-searchlister-search__sortbutton" data-cmp-hook-formssearch="filter">
                 <option value="" selected="selected">Filter By</option>
                 <option value="tags">Tag</option>
                 <option value="title">Title</option>
                 <option value="description">Description</option>
             </select>
         </div>
-        <div class="cmp-searchnlister-search__sortbutton--wrapper" data-sly-test.advancedSearch="${!model.advancedSearchDisabled}">
-            <select class="cmp-searchnlister-search__sortbutton cmp-searchnlister-search__sortbutton--right" data-cmp-hook-formssearch="sort">
+        <div class="cmp-searchlister-search__sortbutton--wrapper" data-sly-test.advancedSearch="${!model.advancedSearchDisabled}">
+            <select class="cmp-searchlister-search__sortbutton cmp-searchlister-search__sortbutton--right" data-cmp-hook-formssearch="sort">
                 <option value="" selected="selected">Sort By</option>
                 <option value="asc">Ascending</option>
                 <option value="desc">Descending</option>
             </select>
         </div>
     </div>
-    <div class="cmp-searchnlister-results cmp-portallister-results" data-cmp-hook-formssearch="results"></div>
-    <div class="cmp-searchnlister__more--wrapper">
-        <div class="cmp-searchnlister__more" data-cmp-hook-formssearch="more">Load more</div>
+    <div class="cmp-searchlister-results cmp-portallister-results" data-cmp-hook-formssearch="results"></div>
+    <div class="cmp-searchlister__more--wrapper">
+        <div class="cmp-searchlister__more" data-cmp-hook-formssearch="more">Load more</div>
     </div>
 
     <script data-cmp-hook-formssearch="itemTemplate"
