+++ ui.tests/test-module/specs/searchnlistercomponent.spec.js
@@ -31,7 +31,7 @@ const sitesSelectors = require('../libs/commons/sitesSelectors'),
 
 describe('Search And Lister - Authoring', function () {
     // we can use these values to log in
-    const   pagePath = "/content/core-components-examples/library/forms-and-communications-portal/searchnlister",
+    const   pagePath = "/content/core-components-examples/library/forms-and-communications-portal/searchlister",
             componentEditPath = pagePath + afConstants.RESPONSIVE_GRID_DEMO_SUFFIX + "/" + afConstants.components.forms.resourceType.fpsnlcomponent.split("/").pop(),
             componentEditPathSelector = "[data-path='" + componentEditPath + "']",
             componentDropPath = pagePath + afConstants.RESPONSIVE_GRID_SUFFIX + "/" + afConstants.components.forms.resourceType.fpsnlcomponent.split("/").pop(),
